+++ composite/MtoULiveLink/unreal/MtoULiveLink/Source/MtoULiveLink/Private/Tests/MtoULiveLinkTestAnimInstance.h
@@ -10,5 +10,6 @@ class UMtoULiveLinkConflictingPostProcess : public UAnimInstance
     GENERATED_BODY()
 
 public:
+    virtual void NativeInitializeAnimation() override;
     virtual void NativeUpdateAnimation(float DeltaSeconds) override;
 };